+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to the "ecosia-search" extension will be documented in this 
 ## [0.2.x]
 
 - Search from clipboard.
-- Add max url lenght setting.
+- Add max url length setting.
 
 ## [0.1.x]